+++ src/store/pricing.cpp
@@ -36,7 +36,7 @@ PRICE price_item(player_type *player_ptr, object_type *o_ptr, int greed, bool fl
     if (price <= 0)
         return 0L;
 
-    int factor = rgold_adj[ot_ptr->owner_race][player_ptr->prace];
+    int factor = rgold_adj[static_cast<int>(ot_ptr->owner_race)][static_cast<int>(player_ptr->prace)];
     factor += adj_chr_gold[player_ptr->stat_index[A_CHR]];
     int adjust;
     if (flip) {
